+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * added wrap_callable_in_context_manager
 
+*