+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.25.3
 
 * CacheDict will now accept time-like strings
+* add list_values_to_indices