+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.25.4
+
+* add JSONAbleDataObject