+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.25.4
 
 * add JSONAbleDataObject
+* improved typing for reraise_as